Remove setting ObservedGeneration attribute logic
Setting the ObservedGeneration attribute on v1.Condition objects is not
properly handled by the SetStatusCondition which lacks the ability to
update the value for existing conditions.
This bug was fixed in
https://github.com/kubernetes/kubernetes/commit/5712e33abcea86e7bf699f40a3a838afc1b7c278
and once it is released in a future stable version this commit can be
reverted to add back the functionality of initializing the
ObservedGeneration attribute using the parent object's Generation.

Signed-off-by: Aurel Canciu <aurelcanciu@gmail.com>