Turn Change into ReconcilateAtChanged predicate
As predicates can be chained and used in conjunction with others,
it is more useful to do just one thing in the predicate: detecting
changes to the meta.ReconcileAtAnnotation.

The generation changes can still be detected by chaining the predicate
together with the predicate.GenerationChangedPredicate from
controller-runtime itself using predicate.Or.

Signed-off-by: Hidde Beydals <hello@hidde.co>