git: drop usage of Name for GitLab operations
During the development of the provider a mistake was made, and the
repository was created using the `Name` field instead of the `Path`.
The result of this is that GitLab will create a `Path` based on the
provided `Name`, with a random (number) suffix added, resulting in
all sorts of issues (and work around attempts) down the line.

By only working with `Path` fields, and assuming that the provided
repository name is a path, the look ups we make are now guaranteed
to always return the same group, subgroup or repository.

Signed-off-by: Hidde Beydals <hello@hidde.co>