Ensure that StartSSH works
The underlying library, gitkit, expects to have a KeyDir supplied in
its *Config when running an SSH server. It's uncertain whether
StartSSH is used anywhere (and it wouldn't work, if it is); but for
the least disruption, I have added a func for supplying the KeyDir
value rather than changing the signature of any existing func.

Since StartSSH blocks when it succeeds, the test has to do a kludge
with a timer to be able to exit.

Signed-off-by: Michael Bridgen <michael@weave.works>