Separate ListenSSH from StartSSH
`StartSSH()` doesn't return until `StopSSH()` is called (or there's an
error). This makes it difficult to use in tests, since you will
usually want to know the address to use before proceeding with the
rest of the test, but the listener must be created before it can be
obtained.

This commit adds a `ListenSSH()`, which creates the SSH server and
listener on localhost, then returns. That means you can call
`ListenSSH()`, get the address, start a goroutine to run StartSSH(),
then complete the test (and `StopSSH()`).

The semantics of how to use SSH server is different now: instead of
StartSSH creating a new server (with a new listener), both ListenSSH()
and StartSSH() create a server _if it hasn't been created
already_. This is necessary because creating a new listener each time
invalidates any address obtained from `SSHAddress()` previously,
making the test control flow above impossible.

Another change, for testing convenience: `SSHAddress()` now returns a
URL you can use with git, like `HTTPAddress()` does.

Signed-off-by: Michael Bridgen <michael@weave.works>