Add method to GitServer to switch auth on
When testing using gittestserver, it's useful to be able to use
authentication. This is especially true when testing against libgit2,
because it expects to use authentication over SSH, and will balk if
it's not requested.

This commit adds a method to *GitServer for switching auth on. There's
no way to switch auth on for just SSH -- it will also be required for
HTTP. For that reason, a username and password are needed.

The credentials are included in the URL returned by
HTTPAddressWithCredentials(), so anything using that will
transparently work whether auth is on or not. If you need to test
authentication with HTTP, use `s.HTTPAddress()` and the credentials
you handed to `s.Auth()`.

Signed-off-by: Michael Bridgen <michael@weave.works>