Add reason and map severity to K8s core event type
This commit adds support for providing a reason to the Event method,
and maps the severity of an event to a K8s core event type. This makes
the emitted events in-line with Kubernetes standards, and is beneficial
to the Flux UI: https://github.com/fluxcd/webui/pull/13

The reason to not fully depend on the K8s event type and drop the
severity is because there have been reports from users wanting to have
access to "debug" events, a type not known to Kubernetes.

Signed-off-by: Hidde Beydals <hello@hidde.co>