Drop ObjectReference parameters and improve logger
This commit addresses the two rough edges mentioned by Michael.

> You have to supply an ObjectReference to both the metrics and the
> events methods.

As the helper methods are always called from within a reconciler, and
both events and metrics must only be recorderd for well known
resources, the methods of the helpers should only be called for objects
that do contain data about the kind and/or version.

This means that calling `reference.GetReference` within the helper
methods will generally not result in any errors, nor be a really
expensive calculation.

In cases where the GVK information for some reason is not available
on the object, `GetReference` will still be capable of falling back to
the manager's `runtime.Scheme` that is now injected to the helper.

Resulting in a - in my personal opinion - fair trade-off, and a much
more friendly API from the consumer's perspective.

> You have to pass logs in to the Events helper.

By relying on the context of the reconcile operation we get a logger
(with object metadata preconfigured) for free.

Signed-off-by: Hidde Beydals <hello@hidde.co>