Add `conditions` helper to `runtime`
This commit adds a helper to work with the conditions of GitOps Toolkit
object kinds, including but not limited to aggregration,
summarization and mirroring of condition types.

The work has been derived from
https://github.com/kubernetes-sigs/cluster-api/tree/7478817225e0a75acb6e14fc7b438231578073d2/util/conditions
but adapted to work with the `metav1.Condition` and
`metav1.ConditionStatus` types.

More concretely, this includes the removal of "condition severity"
related functionalities, as this is not supported by the
`metav1.Condition` type.

The following work is still required before it can be considered ready
for consumption:

* Support for "negative polarity" or "normal-false" conditions; this is
  required to properly support the `Stalled` and `Reconciling`
  conditions of `meta` (and `kstatus`).
* Summarization to other target conditions than `Ready`; results in a
  more generic API.
* Support setting the `ObservedGeneration` in the status condition.

Signed-off-by: Hidde Beydals <hello@hidde.co>