Update K8s dependencies to v1.20.7
This ensures we are no longer affected by GO-2021-0053.

In an initial try an attempt was made to also deal with GO-2020-0018
by updating the `github.com/satori/go.uuid` dependency to
`v1.2.1-0.20181028125025-b2ce2384e17b`.

This is however at present not possible due to a change of the `NewV4`
method signature (single value return -> multiple-value), and the
impact of this should be low (if not non-existent) as the
`gittestserver` package is only used during tests.

Signed-off-by: Hidde Beydals <hello@hidde.co>