Make runtime errors package errors more generic
Quite a lot of the errors in the package had close to the same meaning,
or could be used together with a context defining wrapping error. They
also assumed knowledge of component implementation details at times.

This commit de-duplicates the and removes the errors in these category,
and also properly documents them.

Signed-off-by: Hidde Beydals <hello@hidde.co>