Make it easier to send events without metadata
While validating the runtime package changes, it quickly become
cumbersome to pass `nil` for every event that didn't require any
metadata to be attached.

This commit solves this annoyance by renaming the existing methods to
`EventWithMeta` and `EventWithMetaf`, and reintroducing the `Event` and
`Eventf` methods with signatures that do not take any metadata string
map.

Signed-off-by: Hidde Beydals <hello@hidde.co>