runtime/patch: add integration test using testenv
This was recorded in 723c6bd1e2a4daba07dee8e49cae853665303f15 as a TODO
because the `testenv` package was not yet available.

The integration tests make use of a `Fake` Kubernetes resource object
that for the time being has been made public from a `testdata` package
in `conditions`. In a future iteration we may want to move it to
`pkg/apis` so that it is available for generic usage during e.g.
testing.

Signed-off-by: Hidde Beydals <hello@hidde.co>