Add GitServer method for installing an update hook
For testing what a controller does when the git server balks at an
operation (e.g., pushing a branch), it's useful to be able to install
a service-side git hook in the test git server. This new method is for
installing an update hook (runs for each updated ref), which is the
one I needed; pre-receive (runs before accepting the push) and
post-receive (runs after accepting the push) are trivial adaptations,
should they be needed.

Signed-off-by: Michael Bridgen <michael@weave.works>