runtime/conditions: add kstatus condition helpers
This commit adds getter and setter helpers to `runtime/conditions` to
make it easier to mark resources with kstatus conditions, and observe
their state.

- The `MarkReconciling` and `MarkStalled` functions delete their
  counterpart condition type while marking the resource with the
  kstatus condition, making it easier to adhere to the kstatus
  standards.
- The `IsReady`, `IsStalled` and `IsReconciling` functions take the
  existence of their counterpart condition types into account, ensuring
  conclusive conditional checks.

`MarkReady` was not taken into consideration, it could technically be
added as syntax sugar, but this would first require a discussion about
a counterpart as you likely want to be able to mark something as "not
ready".

To determine a Ready condition at present, the advice is to compose one
using `SetSummary` based on the defined condition types from your own
API, for example:

```go
conditions.SetSummary(
	&obj,
	meta.ReadyCondition,
	conditions.WithConditions(
		foov1.FooCondition,
		foov1.BarCondition,
		foov1.BazCondition,
	),
)
```

Signed-off-by: Hidde Beydals <hello@hidde.co>