Account for reconciling condition in patch example
At present the example of how to use patch.Helper unconditionally
clears the reconciling condition on a "clean" exit (no error or
explicit requeue). This is on the assumption that the condition is
used only within the Reconcile method to indicate that the resource is
being processed.

However, in general it's possible for Reconcile to exit cleanly
without the reconciliation having reached a ready state, but without
definitively failing either. For example, a Deployment would need to
visit Reconcile several times during a rolling upgrade before it has
rolled out a new pod spec.

On the expectation that people may borrow the example, I think it is
better to treat the Reconciling condition as an indication that there
is more work to be done.

Signed-off-by: Michael Bridgen <michael@weave.works>