testenv: Remove klog init flags
Removes klog.InitFlags() because the klog flags aren't parsed before
using klog. To avoid conflicting with other flags, it's better to not
populate the global flagset.
This results in default klogr to be set as the default
controller-runtime logger if not set in the test already. The logger
setup not being in init() allows users to set their own logger before
creating testenv Environment.

Signed-off-by: Sunny <darkowlzz@protonmail.com>