runtime/conditions: Add merge option WithLatestGeneration
Adds merge option WithLatestGeneration() to consider the condition's
observed generation while merging. When withLatestGeneration is set,
the condition groups are created based on the generation of the
conditions. The condition groups with the latest generations are
grouped together and sorted based on their merge priority. The
condition groups with old generations don't contain the up-to-date
information and hence not considered.
The merge behavior remains the same when no generation is specified.

Signed-off-by: Sunny <darkowlzz@protonmail.com>