update KnownKey.Matches to initialize it's own hasher
Previously, KnownKey.Matches() accepted a SHA256 hasher as an argument,
which could lead to unintended bugs when calling it in a loop. This
eliminates that, by initializing a new hasher itself instead of relying
on the caller for the same.
Enables us to fix a regression in the source-controller: https://github.com/fluxcd/image-automation-controller/issues/378

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>