git: introduce `CloneOptions` replacing `CheckoutOptions` with a better API
`CheckoutOptions` is renamed aptly to `CloneOptions` since it configures
cloning behvaior.
A new API `CheckoutStrategy` is introduced and embedded into `CloneOptions`
to better differentiate between options related to cloning and checking
out.
`LastRevision` is renamed to `LastObservedCommit` to make its intended value is
more obvious, and its usage is made clearer with the help of additional docs.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>