git: integrate `WriteFile()` into `Commit()` and add `CommitOptions`
Introduce `CommitOptions` which encapsulates commit configurations such
as an OpenPGP signer and a map of files to be included as a part of the
commit. Introduce `CommitOpton` as a setter for `CommitOptions`.

Remove `WriteFile()` from `RepositoryWriter`. Modify `Commit()` to
accept `CommitOption` as a variadic parameter and integrate the
functionality of `WriteFile()` into `Commit()`.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>