runtime/patch: Add SerialPatcher
SerialPatcher is a wrapper around the patch helper to help perform
consecutive patching of object in a transparent manner without the need
to keep track of the object state after patching.

The SerialPatcher persists the a before state of the object and a
kubernetes client. The patch operations configurations aren't persisted
and are accepted as argument to SerialPatcher.Patch(). This makes the
patcher reusable in different scenarios during its lifetime.

Signed-off-by: Sunny <darkowlzz@protonmail.com>