oci/auth: Fix aws login with aws-sdk-go-v2
As per aws-sdk-go-v2 docs, new config for creating a session should use
config.LoadDefaultConfig() and account ID is not needed for ECR
authorization token.

In order to maintain the auth API which is shared with other provider,
return an empty client for NewClient(). The login implementation loads
the default config if the client config is nil. A test can stub the
client config before calling login.

Signed-off-by: Sunny <darkowlzz@protonmail.com>