go-git: transform revision for last observed tag
This ensures this strategy is backwards compatible as well, as I
accidentally forgot to do this for go-git in #404.

This lacks tests (as does the libgit2 implementation), and it would
probably be good to make them less dynamic as a follow up, as that's
how this slipped through the cracks.

Signed-off-by: Hidde Beydals <hello@hidde.co>