sourceignore: make a copy of domain for pattern
LoadIgnorePatterns() modified the domain string slice after creating
ignore pattern, due to which the domain of certain patterns created
below the first directory level with multiple directories at the same
level resulted in the patterns to have the domain of the last directory
in the same directory level.

This is fixed by making a copy of the domain before creating a pattern.

Signed-off-by: Sunny <darkowlzz@protonmail.com>