actions: make envtest action composite
This drops the `Dockerfile` approach in favor of a composite GitHub
Action.

The `setup-envtest` is installed into the `$RUNNER_TOOL_CACHE`, using
the latest commit made to the `kubernetes-sigs/controller-runtime`
sub-directory which hosts the tool.

In addition, the envtest assets themselves are downloaded to a separate
`$RUNNER_TOOL_CACHE` directory. Allowing them to have a lifecycle
separate from the `setup-envtest` tool itself.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>