Update terraform config to use tags
Update the terraform configurations for all the cloud providers to use
tags. The tags are provided with input variables "tags" which is common
across the cloud providers.

In case of AWS, due to a bug in the aws terraform provider when using
tags with dynamic names, replace the usage of random_pet for random name
generation with a statically set input variable "rand". This is randomly
generated outside of terraform runtime using $RANDOM. The upstream issue
is linked with details in .env.sample file and
fluxcd/test-infra/utils/tags module.

Also update the requirements docs for AWS and GCP with some more
details.

Signed-off-by: Sunny <darkowlzz@protonmail.com>