git: Reimplement cloneRefName using cloneTag's implementation to support annotated tags
`cloneRefName()` failed to support annotated tags and would attempt to
clone the annotated tag object's id as if it were a commit object. This
would cause the following error (with the tag's hash, not the commit):
  unable to resolve commit object for 'da39a3ee5e6b4b0d3255bfef95601890afd80709': object not found

A test for `cloneRefName()` using an annotated tag was added.

`cloneTag()` already had support for annotated tags, and could actually
be modified to support cloning any ref with no modification. The
implementation was moved to `cloneRefName()` and now `cloneTag()` is a
wrapper for `cloneRefName()`.

Signed-off-by: Blake Burkhart <bburky@bburky.com>