Implement new OIDC Login method and update old Login method to ignore ref argument
The Login function in oci/auth no longer uses the name.Reference argument and instead
gets the registry host from the image argument and passes that to the Login method of
the different region. This is to allow login to consumers of this function that don't
have a repository name and just a registry host.

Signed-off-by: Somtochi Onyekwere <somtochionyekwere@gmail.com>