oci: Improve handling of cosgin artifacts
Cosign uses "special" tags that have the following naming format:

<digestAgorithm>-<digestHex>.<artifactType(.att, .sbom, .sig)>

These are used to attach attestations, SBOMs and signatures to a
subject image reference by the first portion of the tag.

Until now Flux would ignore `.sig` tags, as its primary goal is to
filter through app image versions. This change extends the list of
omitted cosing artifacts to cover `.att` and `.sbom`.

In some cases it's desirable to observe cosgin artifacts, e.g. in
order to determine wheather an image had been signed or not.
Hence this change introduces a new field in `ListOptions`.

Co-authored-by: Sunny <darkowlzz@protonmail.com>
Signed-off-by: Ilya Dmitrichenko <errordeveloper@gmail.com>