runtime: add `jitter` utils and options
This adds a `jitter` package to runtime to allow the configuration of a
random +/-% jitter to the interval at which resources are requeued.

Applying jitter to the interval duration can be useful to mitigate
spikes in memory and CPU usage caused by many resources being
configured with the same interval.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>