runtime/metrics: Delete metrics on object delete
Delete the object metrics when the object is deleted. This ensures
that stale metrics about a deleted object is no longer exported.

As a result, the `ConditionDelete` is no longer needed. Another reason
to not have `ConditionDelete` is that a condition can only be one of
True, False or Unknown.

This introduces new delete methods in the low level metrics Recorder. In
the high level controller metrics, a list of owned finalizers is
introduced which is used to determine if an object is being deleted.
The existing Record*() methods are updated to check if the given object
is deleted, and call record or delete based on that. The user of this
API has to pass in the finalizer they write on object they maintain to
the metrics helper and record the metrics at the very end of the
reconciliation so that the final object state can be used to determine
if the metrics can be deleted safely.

To allow creating multiple instances of metrics helper, the metrics
collector registration is now done using a new function in metrics
package called MustMakeRecorder() which returns a metrics.Recorder.
metrics.Recorder can be used to create multiple metrics helpers with
different attributes if required, sharing the same underlying metrics
recorder.

Signed-off-by: Sunny <darkowlzz@protonmail.com>