oci: Use controller-runtime pkg/log explicitly
This helps avoid importing the controller-runtime pkg/client/config
package which has a flag initialization for "kubeconfig". This results
in all the users of the oci package to also have the flag initialized in
their applications.
The usage of controller-runtime in oci package is just for logging.
Importing the pkg/log package specifically helps avoid importing the
client config which sets the flag.

Signed-off-by: Sunny <darkowlzz@protonmail.com>