ssa: add post-dry-run normalize function
To facilitate fixing issues to Kubernetes resources as returned by a
server-side dry-run apply.

The behavior of the function itself equals to the previous behavior
of `fixHorizontalPodAutoscaler`. However, this is now exposed to allow
other consumers (e.g. the to be introduced `jsondiff` sub-library) to
make use of the same patches without further code duplication.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>