ssa: improve data masking
Instead of masking the data with `****` or `*****`, provide an easier
to distinguish `*** (before)` and `*** (after)` for values which have
changed. Matching the output of `kubectl diff`.

In addition, make the function public as it can be of use outside of the
`Diff` of the resource manager, and allow passing only one instance of
an object (for scenarios where a resource is either created or deleted).

Signed-off-by: Hidde Beydals <hidde@hhh.computer>