ssa/jsondiff: add graceful list option
This allows `UnstructuredList` to continue the diff operation when an
error occurs for a single resource. Which may be desirable behavior
when you want to act on any detected divergence, even if a portion of
the operation failed.

The parsing of an `IgnoreRule` itself _does_ continue to immediately
return an error, as otherwise the produced diff could be inaccurate
(instead of being unable to provide one).

Signed-off-by: Hidde Beydals <hidde@hhh.computer>