integration-cleanup: Enable reaper delete
The integration test resources cleanup program, reaper, used to be run
in dry-run mode before to only report on leftover resources that were
more than a day old. Recently, once in a few days, some GKE clusters
would provision in unusable state. In the GKE web console, the cluster
shows the following error:

```
All cluster resources were brought up, but: only 2 nodes out of 3 have
registered; cluster may be unhealthy.
```

And trying to connect to the cluster using cloudshell results in
connection error:

```
WARNING: cluster flux-test-casual-oryx is not RUNNING. The kubernetes
API may or may not be available. Check the cluster status for more
information.
```

The cluster is not usage.
When this happens, terraform waits for the cluster to be ready and
timeout. Due to the timeout, the cluster is not written to the terraform
state file and it can't be deleted by running terraform destroy.

The GKE and individual node logs and monitoring pages don't show any
other details about the issue,

In the past, these resources were manually deleted from the web console.
Since this failure can happen any time, enabling the reaper cleanup took
with retention period 1 hour would be helpful to ensure such unusable
cluster get deleted within a hour. Test clusters don't run for more than
20 minutes. Only the resources with the tag `ci=true` will get deleted.
Also, update the cron time for cleanup to run every hour.

Signed-off-by: Sunny <github@darkowlzz.space>