Add a store interface with default implementations
The store interface is a simplified version of client go store. If an
object can be cached there, it should be in our store implementations.

We also provide a default way to generate keys, but we allow users to
provide their own function. This should enable all needed levels of
isolation.

The provided implementations are an expiring cache and a LRU cache.

The cache is self-instrumented.

Signed-off-by: Soule BA <bah.soule@gmail.com>