workflows/integration-*: set git and oci env vars
With the recent changes to the integration tests, the type of tests
supported by the infrastructure need to be explicitly set.
OCI infrastructure is supported by all of our setup.
Git infrastructure is only set for Azure only. It is ineffective for now
as we don't have an azure account to run the tests at present.

Signed-off-by: Sunny <github@darkowlzz.space>